Move RPC files to rpc directory
