+++ packages/cli/src/rpc/util.ts
